requirements.txt
flask
flask-mongoengine
flask-mqtt
oauth2client
flask-cors
pyjwt

